requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
altair==6.0.0
    # via streamlit
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   loguru
faker==39.0.0
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
idna==3.11
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kafka-python==2.3.0
    # via -r requirements.in
loguru==0.7.3
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
narwhals==2.14.0
    # via altair
numpy==2.4.0
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.0.0
    # via streamlit
protobuf==6.33.2
    # via streamlit
py4j==0.10.9.7
    # via pyspark
pyarrow==22.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pyspark==3.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
redis==7.1.0
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.52.2
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   referencing
    #   streamlit
tzdata==2025.3
    # via
    #   faker
    #   pandas
urllib3==2.6.2
    # via requests
watchdog==6.0.0
    # via streamlit
win32-setctime==1.2.0
    # via loguru
